# rozert-pay/swagger.yml
openapi: 3.0.3
info:
  title: Rozert Pay
  version: 1.0.0
  description: Rozert Payment API
paths:
...
  /api/payment/v1/appex/deposit/:
    post:
      operationId: appex.deposit
      summary: Create Appex deposit transaction
      tags:
      - Appex
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppexDeposit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AppexDeposit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AppexDeposit'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/appex/withdraw/card-data/:
    post:
      operationId: appex.withdraw.card-data
      summary: Create Appex withdrawal transaction by card data
      tags:
      - Appex
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppexCardWithdraw'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AppexCardWithdraw'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AppexCardWithdraw'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/appex/withdraw/card-token/:
    post:
      operationId: appex.withdraw.card-token
      description: |2

        Card token can be obtained from first successfull deposit transaction.
        See card_token parameter in transaction response.
      summary: Create Appex withdrawal transaction by card token.
      tags:
      - Appex
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppexCardTokenWithdraw'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AppexCardTokenWithdraw'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AppexCardTokenWithdraw'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
...
  /api/payment/v1/cardpay-applepay/deposit/:
    post:
      operationId: cardpay_applepay.deposit
      summary: Create cardpay applepay deposit transaction
      tags:
      - Cardpay Applepay
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardpayApplepayDeposit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CardpayApplepayDeposit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CardpayApplepayDeposit'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
...
  /api/payment/v1/cardpay-applepay/withdraw/:
    post:
      operationId: cardpay_applepay.withdraw
      summary: Create Cardpay Applepay withdrawal transaction
      tags:
      - Cardpay Applepay
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardpayApplepayWithdraw'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CardpayApplepayWithdraw'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CardpayApplepayWithdraw'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/cardpay-cards/deposit/:
    post:
      operationId: cardpay_cards.deposit
      summary: Create deposit transaction
      tags:
      - Cardpay Bankcard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardpayBankcardDeposit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CardpayBankcardDeposit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CardpayBankcardDeposit'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/cardpay-cards/deposit_card_token/:
    post:
      operationId: cardpay_cards.deposit_2
      summary: Create deposit transaction using card token from previous deposits
      tags:
      - Cardpay Bankcard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardpayBankcardDepositByToken'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CardpayBankcardDepositByToken'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CardpayBankcardDepositByToken'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/cardpay-cards/withdraw/card-data/:
    post:
      operationId: cardpay_cards.withdraw.card-data
      summary: Create Cardpay Bankcard withdrawal transaction by card data
      tags:
      - Cardpay Bankcard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardpayBankcardWithdraw'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CardpayBankcardWithdraw'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CardpayBankcardWithdraw'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/cardpay-cards/withdraw/card-token/:
    post:
      operationId: cardpay_cards.withdraw.card-token
      description: "\nCard token can be obtained from first successfull deposit transaction.\n\
        See card_token parameter in transaction response.\n    "
      summary: Create Cardpay Bankcard withdrawal transaction by card token.
      tags:
      - Cardpay Bankcard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardpayBankcardWithdrawByToken'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CardpayBankcardWithdrawByToken'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CardpayBankcardWithdrawByToken'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
...
  /api/payment/v1/d24-mercadopago/deposit/:
    post:
      operationId: Create deposit transaction
      tags:
      - D24 MercadoPago
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/D24MercadoPagoDeposit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/D24MercadoPagoDeposit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/D24MercadoPagoDeposit'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          description: No response body
...
  /api/payment/v1/ilixium/deposit/:
    post:
      operationId: ilixium.deposit
      summary: Create Ilixium deposit transaction
      tags:
      - Ilixium
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IlixiumDeposit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IlixiumDeposit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IlixiumDeposit'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/ilixium/withdraw/:
    post:
      operationId: ilixium.withdraw
      summary: Create Ilixium withdrawal transaction
      tags:
      - Ilixium
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IlixiumWithdraw'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/IlixiumWithdraw'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IlixiumWithdraw'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
...
  /api/payment/v1/nuvei/deposit/:
    post:
      operationId: nuvei.deposit
      summary: Create Nuvei deposit transaction
      tags:
      - Nuvei
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuveiDeposit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NuveiDeposit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NuveiDeposit'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
  /api/payment/v1/nuvei/withdraw/card-data/:
    post:
      operationId: nuvei.withdraw.card-data
      summary: Create Nuvei withdrawal transaction by card data
      tags:
      - Nuvei
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuveiWithdraw'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NuveiWithdraw'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NuveiWithdraw'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
...
  /api/payment/v1/worldpay/ddc-jwt/:
    get:
      operationId: worldpay.ddc_jwt
      summary: Generate JWT for DDC (Device Data Collection) step
      tags:
      - Worldpay
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DDCJwtResponse'
          description: ''
  /api/payment/v1/worldpay/deposit/:
    post:
      operationId: worldpay.deposit
      summary: Create Worldpay deposit transaction
      tags:
      - Worldpay
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorldpayDeposit'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/WorldpayDeposit'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/WorldpayDeposit'
        required: true
      security:
      - HMACAuthentication: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
          description: ''
components:
  schemas:
...
    AppexCardTokenWithdraw:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        user_data:
          allOf:
          - $ref: '#/components/schemas/UserDataSerializerOptional'
          nullable: true
        card:
          $ref: '#/components/schemas/_CardTokenSerialzer'
      required:
      - amount
      - card
      - currency
      - customer_id
      - wallet_id
    AppexCardWithdraw:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        user_data:
          allOf:
          - $ref: '#/components/schemas/UserDataSerializerOptional'
          nullable: true
        card:
          allOf:
          - $ref: '#/components/schemas/_CardNoCVV'
          description: Card details without CVV
      required:
      - amount
      - card
      - currency
      - customer_id
      - wallet_id
    AppexDeposit:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
          description: Wallet ID. Will be provided by rozert.
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        card:
          allOf:
          - $ref: '#/components/schemas/_Card'
          description: Card details
        user_data:
          $ref: '#/components/schemas/AppexUserData'
      required:
      - amount
      - card
      - currency
      - customer_id
      - user_data
      - wallet_id
...
    CardpayApplepayDeposit:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
          description: Wallet ID. Will be provided by rozert.
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        user_data:
          $ref: '#/components/schemas/UserData'
        encrypted_data:
          type: string
          description: This data should be passed from client side
      required:
      - amount
      - currency
      - customer_id
      - encrypted_data
      - user_data
      - wallet_id
...
    CardpayApplepayWithdraw:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        user_data:
          allOf:
          - $ref: '#/components/schemas/UserDataSerializerOptional'
          nullable: true
        encrypted_data:
          type: string
          description: This data should be passed from client side
      required:
      - amount
      - currency
      - customer_id
      - encrypted_data
      - wallet_id
    CardpayBankcardDeposit:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
          description: Wallet ID. Will be provided by rozert.
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        card:
          allOf:
          - $ref: '#/components/schemas/_Card'
          description: Card details
        user_data:
          $ref: '#/components/schemas/CardpayBankcardWithdrawUserData'
      required:
      - amount
      - card
      - currency
      - customer_id
      - user_data
      - wallet_id
    CardpayBankcardDepositByToken:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
          description: Wallet ID. Will be provided by rozert.
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        card:
          $ref: '#/components/schemas/_CardTokenSerialzer'
        user_data:
          $ref: '#/components/schemas/CardpayBankcardWithdrawUserData'
      required:
      - amount
      - card
      - currency
      - customer_id
      - user_data
      - wallet_id
    CardpayBankcardWithdraw:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        user_data:
          $ref: '#/components/schemas/CardpayBankcardWithdrawUserData'
        card:
          allOf:
          - $ref: '#/components/schemas/_CardNoCVV'
          description: Card details without CVV
      required:
      - amount
      - card
      - currency
      - customer_id
      - user_data
      - wallet_id
    CardpayBankcardWithdrawByToken:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        customer_id:
          type: string
          description: Customer unique ID
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        currency:
          type: string
          maxLength: 3
        redirect_url:
          type: string
          format: uri
          nullable: true
          description: Redirect URL for payment system to redirect user after payment.
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback url for payment system
        user_data:
          allOf:
          - $ref: '#/components/schemas/UserDataSerializerOptional'
          nullable: true
        card:
          $ref: '#/components/schemas/_CardTokenSerialzer'
      required:
      - amount
      - card
      - currency
      - customer_id
      - wallet_id
...
    DDCJwtResponse:
      type: object
      properties:
        jwt_token:
          type: string
          description: DDC JWT token for device data collection
      required:
      - jwt_token
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/TransactionResponseStatusEnum'
        decline_code:
          type: string
        decline_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        instruction:
          allOf:
          - $ref: '#/components/schemas/Instruction'
          nullable: true
          description: 'Instruction for customer. Required for deposits for: paycash'
        callback_url:
          type: string
          format: uri
          nullable: true
          description: Callback URL for payment system to notify about transaction
            status change.
        customer_id:
          type: string
          nullable: true
          description: Internal customer ID
        external_customer_id:
          type: string
          nullable: true
          description: External customer ID (provided by merchant)
        type:
          $ref: '#/components/schemas/TransactionResponseTypeEnum'
        currency:
          type: string
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        form:
          allOf:
          - $ref: '#/components/schemas/FormData'
          nullable: true
          description: Form data for redirecting user to payment system. If presented,
            merchant customer should be redirected or form data must be submitted.
        external_account_id:
          type: string
          nullable: true
          description: External account of user performed deposit. Payment system
            specific.
        user_data:
          allOf:
          - $ref: '#/components/schemas/UserData'
          nullable: true
        card_token:
          type: string
          nullable: true
          description: For card payment systems, this is the token of the card used
            for deposit.You should store this token and use it for withdrawals when
            needed.
      required:
      - amount
      - created_at
      - currency
      - decline_code
      - decline_reason
      - id
      - status
      - type
      - updated_at
...
    TransactionResponseTypeEnum:
      enum:
      - deposit
      - withdrawal
      type: string
      description: |-
        * `deposit` - Deposit
        * `withdrawal` - Withdrawal
...
  securitySchemes:
    BetmasterSecretKeyAuthentication:
      type: apiKey
      in: header
      name: X-Back-Secret-Key
    CSRFExemptSessionAuthentication: []
    HMACAuthentication:
    - type: apiKey
      in: header
      name: X-Merchant-Id
    - type: apiKey
      in: header
      name: X-Signature
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid